Blueprints refactor
