+++ elementalcms/presenter/__init__.py
@@ -1,0 +1,1 @@
+from .presenter import presenter
